@@ -4,7 +4,7 @@
   "author": "Ultimate Gadget Laboratories",
   "main": "electron/dist/electron-main.js",
   "version": "1.2.13",
-  "firmwareVersion": "8.5.3",
+  "firmwareVersion": "8.5.4",
   "deviceProtocolVersion": "4.4.0",
   "userConfigVersion": "4.0.1",
   "hardwareConfigVersion": "1.0.0",
@@ -15,22 +15,25 @@
   },
   "license": "GPL-3.0",
   "engines": {
-    "node": ">=8.12.0 <9.0.0",
-    "npm": ">=6.4.0 <7.0.0"
+    "node": ">=12.0.0 <13.0.0",
+    "npm": ">=6.9.0 <7.0.0"
   },
   "devDependencies": {
+    "@types/decompress": "4.2.0",
     "@types/electron-devtools-installer": "2.0.2",
     "@types/electron-settings": "3.0.0",
-    "@types/fs-extra": "5.0.1",
-    "@types/jasmine": "2.8.8",
+    "@types/file-saver": "0.0.1",
+    "@types/fs-extra": "8.0.0",
+    "@types/jasmine": "3.3.12",
+    "@types/jasminewd2": "2.0.3",
     "@types/jquery": "3.3.29",
     "@types/jsonfile": "4.0.1",
-    "@types/lodash": "4.14.121",
+    "@types/lodash": "4.14.136",
     "@types/node": "8.0.53",
     "@types/node-hid": "0.7.0",
     "@types/request": "2.0.8",
     "@types/semver": "5.5.0",
-    "@types/usb": "1.1.3",
+    "@types/tmp": "0.0.33",
     "autoprefixer": "6.5.3",
     "buffer": "5.0.6",
     "check-node-version": "^3.2.0",
@@ -41,28 +44,27 @@
     "decompress": "4.2.0",
     "decompress-tarbz2": "4.1.1",
     "devtron": "1.4.0",
-    "electron": "2.0.16",
+    "electron": "5.0.9",
     "electron-builder": "20.34.0",
     "electron-debug": "1.5.0",
     "electron-devtools-installer": "2.2.3",
     "electron-log": "2.2.16",
-    "electron-rebuild": "1.8.2",
+    "electron-rebuild": "1.8.5",
     "electron-settings": "3.1.4",
     "electron-updater": "2.21.4",
     "exports-loader": "0.6.3",
     "file-loader": "0.10.0",
-    "fs-extra": "5.0.0",
+    "fs-extra": "8.1.0",
     "gh-pages": "2.0.1",
     "html-webpack-plugin": "3.2.0",
-    "jasmine": "2.8.0",
-    "jasmine-core": "2.8.0",
-    "jasmine-node": "2.0.1",
+    "jasmine": "3.4.0",
+    "jasmine-core": "3.4.0",
+    "jasmine-node": "3.0.0",
     "jasmine-ts": "0.3.0",
     "jsonfile": "4.0.0",
-    "lerna": "3.13.0",
-    "lodash": "4.17.11",
-    "mkdirp": "0.5.1",
-    "node-hid": "0.7.3",
+    "lerna": "3.16.4",
+    "lodash": "4.17.15",
+    "node-hid": "0.7.9",
     "npm-run-all": "4.0.2",
     "nrf-intel-hex": "1.3.0",
     "postcss-url": "8.0.0",
@@ -70,12 +72,12 @@
     "request": "2.88.0",
     "rimraf": "2.6.1",
     "source-map-support": "0.5.9",
-    "stylelint": "9.6.0",
+    "stylelint": "10.0.0",
     "svg-sprite": "1.5.0",
     "ts-loader": "5.3.3",
-    "ts-node": "7.0.1",
-    "tslint": "5.13.0",
-    "typescript": "3.3.3",
+    "ts-node": "8.3.0",
+    "tslint": "5.18.0",
+    "typescript": "3.5.3",
     "webpack": "4.29.5",
     "webpack-cli": "3.2.3"
   },
@@ -100,6 +102,5 @@
     "clean": "lerna exec rimraf ./node_modules ./dist && rimraf ./node_modules ./dist ./tmp",
     "predeploy-gh-pages": "lerna run build:web --scope=uhk-web",
     "deploy-gh-pages": "gh-pages -d packages/uhk-web/dist"
-  },
-  "dependencies": {}
+  }
 }
