@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, Output, SimpleChanges, ChangeDetectionStrategy } from '@angular/core';
 import { animate, state, trigger, style, transition } from '@angular/animations';
 import { DomSanitizer, SafeStyle } from '@angular/platform-browser';
-import { Module } from 'uhk-common';
+import { HalvesInfo, Module } from 'uhk-common';
 
 import { SvgModule } from '../module';
 import { SvgModuleProviderService } from '../../../services/svg-module-provider.service';
@@ -23,12 +23,22 @@ import { LastEditedKey } from '../../../models';
     animations: [
         trigger('split', [
             state('rotateLeft', style({
-                transform: 'translate(-3%, 15%) rotate(4deg) scale(0.92, 0.92)'
+                transform: 'translate(2%, 30%) rotate(10.8deg) scale(0.80, 0.80)'
             })),
             state('rotateRight', style({
-                transform: 'translate(3%, 15%) rotate(-4deg) scale(0.92, 0.92)'
+                transform: 'translate(-2%, 30.7%) rotate(-10deg) scale(0.80, 0.80)'
             })),
             transition('* <=> *', animate(500))
+        ]),
+        trigger('fadeKeyboard', [
+            state('visible', style({
+                opacity: 1
+            })),
+            state('invisible', style({
+                opacity: 0
+            })),
+            transition('visible => invisible', animate(500)),
+            transition('invisible => visible', animate(500))
         ]),
         trigger('fadeSeparator', [
             state('visible', style({
@@ -37,8 +47,18 @@ import { LastEditedKey } from '../../../models';
             state('invisible', style({
                 opacity: 0
             })),
-            transition('visible => invisible', animate(500)),
-            transition('invisible => visible', animate(1500))
+            transition('visible => invisible', animate('200ms')),
+            transition('invisible => visible', animate('200ms 500ms'))
+        ]),
+        trigger('moveDescription', [
+            state('down', style({
+                'margin-top': '0.5em'
+            })),
+            state('up', style({
+                'margin-top': '-6.5%'
+            })),
+            transition('down => up', animate(500)),
+            transition('up => down', animate(500))
         ])
     ]
 })
@@ -47,7 +67,7 @@ export class SvgKeyboardComponent {
     @Input() capturingEnabled: boolean;
     @Input() selectedKey: { layerId: number, moduleId: number, keyId: number };
     @Input() selected: boolean;
-    @Input() halvesSplit: boolean;
+    @Input() halvesInfo: HalvesInfo;
     @Input() keyboardLayout = KeyboardLayout.ANSI;
     @Input() description: string;
     @Input() showDescription = false;
@@ -60,16 +80,18 @@ export class SvgKeyboardComponent {
     modules: SvgModule[];
     viewBox: string;
     moduleAnimationStates: string[];
+    moduleVisibilityAnimationStates: string[];
     separator: SvgSeparator;
     separatorStyle: SafeStyle;
     separatorAnimation = 'visible';
+    descriptionAnimation = 'down';
 
     constructor(private svgModuleProvider: SvgModuleProviderService,
                 private sanitizer: DomSanitizer) {
         this.modules = [];
         this.viewBox = '-520 582 1100 470';
-        this.halvesSplit = false;
         this.moduleAnimationStates = [];
+        this.moduleVisibilityAnimationStates = [];
     }
 
     ngOnInit() {
@@ -77,7 +99,7 @@ export class SvgKeyboardComponent {
     }
 
     ngOnChanges(changes: SimpleChanges) {
-        if (changes.halvesSplit) {
+        if (changes.halvesInfo) {
             this.updateModuleAnimationStates();
         }
 
@@ -110,12 +132,21 @@ export class SvgKeyboardComponent {
     }
 
     private updateModuleAnimationStates() {
-        if (this.halvesSplit) {
-            this.moduleAnimationStates = ['rotateRight', 'rotateLeft'];
-            this.separatorAnimation = 'invisible';
-        } else {
+        if (this.halvesInfo.areHalvesMerged) {
             this.moduleAnimationStates = [];
             this.separatorAnimation = 'visible';
+            this.descriptionAnimation = 'down';
+        } else {
+            this.moduleAnimationStates = ['rotateRight', 'rotateLeft'];
+            this.separatorAnimation = 'invisible';
+            this.descriptionAnimation = 'up';
+        }
+
+        if (this.halvesInfo.isLeftHalfConnected) {
+            this.moduleVisibilityAnimationStates = ['visible', 'visible'];
+        } else {
+            this.moduleVisibilityAnimationStates = ['visible', 'invisible'];
+
         }
     }
 
